+++ tests/testthat/tests.support.R
@@ -13,8 +13,13 @@ test_that("Test that brm.links return an error on wrong links", {
   expect_error(brm.link(c("categorical","probit")), "probit is not a valid link for family categorical")
 })
 
-test_that("Test that extract.effects finds all variables", {
-  expect_equal(extract.effects(t2_brand_recall ~ psi_expsi + psi_api_probsolv + 
-                               psi_api_ident + psi_api_intere + psi_api_groupint)$all, 
-               t2_brand_recall ~ psi_expsi + psi_api_probsolv + psi_api_ident + psi_api_intere + psi_api_groupint)
+
+test_that("Test that is.formula is TRUE for formulas and otherwise FALSE", {
+  expect_equal(is.formula(y~1), TRUE)
+  expect_equal(is.formula("a"), FALSE)
+  expect_equal(is.formula(list(y~1, ~1)), TRUE)
+  expect_equal(is.formula(list(y~1,1)), TRUE)
+  expect_equal(is.formula(list("a",1)), FALSE)
+  expect_equal(is.formula(list(y~1, ~1), or = FALSE), TRUE)
+  expect_equal(is.formula(list(y~1,1), or = FALSE), FALSE)
 })